+++ DataManager/include/hpd_configuration.h
@@ -23,37 +23,31 @@
   The views and conclusions contained in the software and documentation are those of the
   authors and should not be interpreted as representing official policies, either expressed*/
 
-/**
- * @file hpd_configuration.h
- * @brief  Methods for managing the configuration structure
- * @author Thibaut Le Guilly
- */
-
 #ifndef CONFIGURATION_H
 #define CONFIGURATION_H
 
-#include "hpd_adapter.h"
+#include <mxml.h>
+#include <jansson.h>
 
-#define ADAPTER_ID_SIZE 2
+typedef struct Configuration Configuration;
+typedef struct Adapter       Adapter;
 
 /**
  * The structure Configuration containing all the Attributes that a Configuration possesses
  */
-typedef struct Configuration Configuration;
-
 struct Configuration
 {
-  AdapterElement *adapter_head;
+   // Navigational members
+   Adapter *adapter_head;
 };
 
-Configuration* 	configurationNew();
-void 		configurationFree(Configuration *config);
-int 		configurationAddAdapter(Configuration *config, Adapter *adapter);
-int 		configurationRemoveAdapter( Configuration *configuration, Adapter *adapter );
-Adapter* 	findAdapter(Configuration *configuration, char *adapter_id);
-mxml_node_t* 	configurationToXml(Configuration *configuration, mxml_node_t *parent);
-json_t* 	configurationToJson(Configuration *configuration);
-char* 		confGenerateDeviceId(Configuration *conf);
+Configuration *configurationNew();
+void           configurationFree(Configuration *config);
+int            configurationAddAdapter(Configuration *config, Adapter *adapter);
+int            configurationRemoveAdapter( Configuration *configuration, Adapter *adapter );
+Adapter       *configurationFindAdapter(Configuration *configuration, char *adapter_id);
+mxml_node_t   *configurationToXml(Configuration *configuration, mxml_node_t *parent);
+json_t        *configurationToJson(Configuration *configuration);
 
 #endif
