+++ DataManager/src/hpd_configuration.c
@@ -31,10 +31,11 @@
 
 
 #include "hpd_configuration.h"
+#include "hpd_adapter.h"
 #include "hp_macros.h"
-
-char *confGenerateAdapterId(Configuration *conf);
-AdapterElement* findAdapterElement(Configuration *configuration, char *adapter_id);
+#include "hpd_error.h"
+#include "utlist.h"
+#include "idgen.h"
 
 Configuration*
 configurationNew()
@@ -47,6 +48,7 @@ cleanup:
   return NULL;
 }
 
+
 void
 configurationFree(Configuration *config)
 {
@@ -54,11 +56,11 @@ configurationFree(Configuration *config)
   {
     if( config->adapter_head != NULL )
     {
-      AdapterElement *iterator, *tmp;
+      Adapter *iterator, *tmp;
       DL_FOREACH_SAFE( config->adapter_head, iterator, tmp )
       {
 	DL_DELETE(config->adapter_head, iterator);
-	adapterElementFree(iterator);
+   iterator->configuration = NULL;
       }
     }
     free_pointer(config);
@@ -71,23 +73,14 @@ configurationAddAdapter(Configuration *configuration, Adapter *adapter)
 {
   if(configuration == NULL || adapter == NULL) return HPD_E_NULL_POINTER;
 
-  char *adapterId = confGenerateAdapterId( configuration );
-  if( adapterId == NULL )
-    return HPD_E_NULL_POINTER;
-
-  adapterSetId( adapter, adapterId );
-
-  AdapterElement *adapterElement = NULL;
-
-  adapterElement = adapterElementNew( adapter );
-  if( adapterElement == NULL )
-  {
-    adapterSetId( adapter, NULL );
-    free(adapterId);
-    return HPD_E_NULL_POINTER;
+  adapter->configuration = configuration;
+  int stat = adapterGenerateId(adapter);
+  if (stat) {
+     adapter->configuration = NULL;
+     return stat;
   }
 
-  DL_APPEND( configuration->adapter_head, adapterElement);
+  DL_APPEND( configuration->adapter_head, adapter);
 
   return HPD_E_SUCCESS;
 }
@@ -98,49 +91,28 @@ configurationRemoveAdapter( Configuration *configuration, Adapter *adapter )
 
   if( configuration == NULL || adapter == NULL ) return HPD_E_NULL_POINTER;
 
-  AdapterElement *adapterElement = findAdapterElement(configuration, adapter->id);
-  if(adapterElement == NULL)
-  {
-    return -1;
-  }
-
-  DL_DELETE(configuration->adapter_head, adapterElement);
-  adapterElementFree(adapterElement);
+  DL_DELETE(configuration->adapter_head, adapter);
+  adapter->configuration = NULL;
 
   return HPD_E_SUCCESS;
 }
 
 
-AdapterElement*
-findAdapterElement(Configuration *configuration, char *adapter_id)
+Adapter*
+configurationFindAdapter(Configuration *configuration, char *adapter_id)
 {
   if( configuration== NULL || adapter_id == NULL ) return NULL;
 
-  AdapterElement *iterator = NULL;
+  Adapter *iterator = NULL;
 
   DL_FOREACH( configuration->adapter_head, iterator )
   {
-    if( strcmp ( adapter_id, iterator->adapter->id ) == 0 )
+    if( strcmp ( adapter_id, iterator->id ) == 0 )
     {
       return iterator;
     }			
   }
   
-  return NULL;
-}
-
-Adapter*
-findAdapter(Configuration *configuration, char *adapter_id)
-{
-  if(configuration == NULL || adapter_id == NULL ) return NULL;
-
-  AdapterElement *adapterElement;
-
-  if( ( adapterElement = findAdapterElement(configuration, adapter_id) ) )
-  {
-    return adapterElement->adapter;
-  }
-
   return NULL;
 
 }
@@ -152,11 +124,11 @@ configurationToXml(Configuration *configuration, mxml_node_t *parent)
 
   configXml = mxmlNewElement(parent, "configuration");
 
-  AdapterElement *iterator;
+  Adapter *iterator;
 
   DL_FOREACH(configuration->adapter_head, iterator)
   {
-    adapterToXml(iterator->adapter, configXml);
+    adapterToXml(iterator, configXml);
   }
 
   return configXml;
@@ -174,7 +146,7 @@ configurationToJson(Configuration *configuration)
     goto error;
   }
 
-  AdapterElement *iterator;
+  Adapter *iterator;
 
   if( ( adapterArray = json_array() ) == NULL )
   {
@@ -183,7 +155,7 @@ configurationToJson(Configuration *configuration)
 
   DL_FOREACH(configuration->adapter_head, iterator)
   {
-    adapter = adapterToJson(iterator->adapter);
+    adapter = adapterToJson(iterator);
     if( ( adapter == NULL ) || ( json_array_append_new(adapterArray, adapter) != 0 ) )
     {
       goto error;
@@ -203,40 +175,4 @@ error:
   return NULL;
 }
 
-char *confGenerateAdapterId(Configuration *conf)
-{
-  char *adapter_id = (char*)malloc((ADAPTER_ID_SIZE+1)*sizeof(char));
-  if( adapter_id == NULL )
-    return NULL;
-  do{
-    rand_str(adapter_id, ADAPTER_ID_SIZE);
-  }while(findAdapter(conf, adapter_id) != NULL);
 
-  return adapter_id;
-}
-
-int
-confDeviceIdExists(Configuration *conf, char *deviceId)
-{
-  AdapterElement *iterator = NULL;
-  DL_FOREACH( conf->adapter_head, iterator )
-  {
-   if(findDevice(iterator->adapter, deviceId) != NULL)
-   {
-     return 1;
-   }
-  }
-  return 0;
-}
-
-char *confGenerateDeviceId(Configuration *conf)
-{
-  char *deviceId = malloc((DEVICE_ID_SIZE+1)*sizeof(char));
-  if( deviceId == NULL )
-    return NULL;
-  do{
-    rand_str(deviceId, DEVICE_ID_SIZE);
-  }while(confDeviceIdExists(conf, deviceId) != 0);
-
-  return deviceId;
-}
