+++ DataManager/src/hpd_device.c
@@ -31,9 +31,15 @@
  */
 
 #include "hpd_device.h"
+#include "hpd_service.h"
+#include "hpd_adapter.h"
+#include "hpd_configuration.h"
 #include "hp_macros.h"
+#include "hpd_error.h"
+#include "utlist.h"
+#include "idgen.h"
 
-char *deviceGenerateServiceId(Device *device);
+#define DEVICE_ID_SIZE 4
 
 /**
  * Create the structure Device with all its parameters
@@ -72,7 +78,8 @@ deviceNew(
     const char *productId,
     const char *version,
     const char *location,
-    const char *type)
+    const char *type,
+    void * data)
 {
   Device *device;
 
@@ -88,6 +95,9 @@ deviceNew(
   null_nok_string_copy(device->type, type);
 
   device->service_head = NULL;
+  device->adapter = NULL;
+
+  device->data = data;
 
   return device;
 
@@ -124,11 +134,11 @@ deviceFree( Device *device )
 
     if( device->service_head )
     {
-      ServiceElement *iterator = NULL, *tmp = NULL;
+      Service *iterator = NULL, *tmp = NULL;
       DL_FOREACH_SAFE( device->service_head, iterator, tmp)
       {
 	DL_DELETE( device->service_head, iterator );
-	serviceElementFree(iterator);
+   iterator->device = NULL;
       }
     }
 
@@ -153,50 +163,19 @@ deviceAddService( Device *device, Service *service )
     return HPD_E_NULL_POINTER;
 
   char *deviceId = NULL;
-  
-  deviceId = deviceGenerateServiceId( device );
-  if( deviceId == NULL )
-    return HPD_E_MALLOC_ERROR;
 
-  serviceSetId( service, deviceId );
-
-  ServiceElement *serviceElement;
-
-  serviceElement = serviceElementNew( service );
-  if( serviceElement == NULL )
-  {
-    serviceSetId( service, NULL );
-    free(deviceId);
-    return HPD_E_MALLOC_ERROR;
-  }
-
-  DL_APPEND( device->service_head, serviceElement);
   service->device = device;
+  int stat;
+  if ((stat = serviceGenerateId(service))) {
+     service->device = NULL;
+     return stat;
+  }
+  
+  service->id = deviceId;
+
+  DL_APPEND( device->service_head, service);
 
   return HPD_E_SUCCESS;
-}
-
-static int removeService(Device *device, Service *service)
-{
-  if( service == NULL || device == NULL ) return HPD_E_NULL_POINTER;
-
-  ServiceElement *iterator, *tmp;
-
-  service->device = NULL;
-
-  DL_FOREACH_SAFE( device->service_head, iterator, tmp )
-  {
-    if( strcmp( service->type, iterator->service->type ) == 0
-	&& strcmp ( service->id, iterator->service->id ) == 0 )
-    {
-      DL_DELETE( device->service_head, iterator );
-      serviceElementFree( iterator );
-      iterator = NULL;
-      return HPD_E_SUCCESS; 
-    }			
-  }
-
-  return -1;
 }
 
 /**
@@ -209,48 +188,45 @@ static int removeService(Device *device, Service *service)
  * @return returns A HPD error code
  */
 int 
-deviceRemoveService( Device *device, Service *service )
-{
-   return removeService(device, service);
-}
-
-int 
-deviceRemoveService2( Service *service )
+deviceRemoveService( Service *service )
 {
    Device *device = service->device;
-   return removeService(device, service);
+   if( service == NULL || device == NULL ) return HPD_E_NULL_POINTER;
+
+   Service *iterator, *tmp;
+
+   service->device = NULL;
+
+   DL_FOREACH_SAFE( device->service_head, iterator, tmp )
+   {
+     if( strcmp( service->type, iterator->type ) == 0
+    && strcmp ( service->id, iterator->id ) == 0 )
+     {
+       DL_DELETE( device->service_head, iterator );
+       iterator->device = NULL;
+       iterator = NULL;
+       return HPD_E_SUCCESS; 
+     }			
+   }
+
+   return -1;
 }
 
-ServiceElement*
-findServiceElement(Device *device, char *service_id)
+Service*
+deviceFindService(Device *device, char *service_id)
 {
   if(device == NULL || service_id == NULL ) return NULL;
 
-  ServiceElement *iterator;
+  Service *iterator;
 
   DL_FOREACH( device->service_head, iterator )
   {
-    if( strcmp ( service_id, iterator->service->id ) == 0 )
+    if( strcmp ( service_id, iterator->id ) == 0 )
     {
       return iterator;
     }			
   }
   
-  return NULL;
-}
-
-Service*
-findService(Device *device, char *service_id)
-{
-  if(device == NULL || service_id == NULL ) return NULL;
-
-  ServiceElement *serviceElement;
-
-  if( ( serviceElement = findServiceElement(device, service_id) ) )
-  {
-    return serviceElement->service;
-  }
-
   return NULL;
 
 }
@@ -271,11 +247,11 @@ deviceToXml(Device *device, mxml_node_t *parent)
   if(device->location != NULL) mxmlElementSetAttr(deviceXml, "location", device->location);
   if(device->type != NULL) mxmlElementSetAttr(deviceXml, "type", device->type);
 
-  ServiceElement *iterator;
+  Service *iterator;
 
   DL_FOREACH( device->service_head, iterator )
   {
-    serviceToXml(iterator->service, deviceXml);
+    serviceToXml(iterator, deviceXml);
   }
 
   return deviceXml;
@@ -342,7 +318,7 @@ deviceToJson(Device *device)
     }
   }
 
-  ServiceElement *iterator;
+  Service *iterator;
 
   if( ( serviceArray = json_array() ) == NULL )
   {
@@ -351,7 +327,7 @@ deviceToJson(Device *device)
 
   DL_FOREACH( device->service_head, iterator )
   {
-    if( json_array_append_new(serviceArray, serviceToJson(iterator->service)) != 0 )
+    if( json_array_append_new(serviceArray, serviceToJson(iterator)) != 0 )
     {
       return NULL;
     }
@@ -363,51 +339,37 @@ deviceToJson(Device *device)
   }
 
   return deviceJson;
-error:
-  if(value) json_decref(value);
-  if(serviceArray) json_decref(serviceArray);
-  if(deviceJson) json_decref(deviceJson);
+//error:
+//  if(value) json_decref(value);
+//  if(serviceArray) json_decref(serviceArray);
+//  if(deviceJson) json_decref(deviceJson);
 }
 
-DeviceElement* 
-deviceElementNew( Device *device )
+static int
+idExists(Configuration *conf, char *deviceId)
 {
-  DeviceElement *to_create;
-
-  if( !device )
-    return NULL;
-
-  to_create = (DeviceElement*)malloc(sizeof(DeviceElement));
-  if( !to_create )
-    return NULL;
-
-  to_create->device = device;
-
-  to_create->next = NULL;
-  to_create->prev = NULL;
-
-  return to_create;
+  Adapter *iterator = NULL;
+  DL_FOREACH( conf->adapter_head, iterator )
+  {
+   if(adapterFindDevice(iterator, deviceId) != NULL)
+   {
+     return 1;
+   }
+  }
+  return 0;
 }
 
-void 
-deviceElementFree( DeviceElement *deviceElement )
+int deviceGenerateId(Device *device)
 {
-  free_pointer(deviceElement);
+   Configuration *conf = device->adapter->configuration;
+
+   char *deviceId = malloc((DEVICE_ID_SIZE+1)*sizeof(char));
+   if (!deviceId) return HPD_E_MALLOC_ERROR;
+   do{
+     rand_str(deviceId, DEVICE_ID_SIZE);
+   }while(idExists(conf, deviceId) != 0);
+
+   device->id = deviceId;
+  return HPD_E_SUCCESS;
 }
 
-char*
-deviceGenerateServiceId(Device *device)
-{
-  char *service_id = (char*)malloc((SERVICE_ID_SIZE+1)*sizeof(char));
-  do{
-    rand_str(service_id, SERVICE_ID_SIZE);
-  }while(findService(device, service_id) != NULL);
-
-  return service_id;
-}
-
-void
-deviceSetId( Device *device, char *id )
-{
-  device->id = id;
-}
