+++ OperationHandler/REST/webserver/src/webserver.c
@@ -82,14 +82,14 @@ void ws_print(struct ws *ws)
    } else {
       printf("   Port: %i\n", ws->settings.port);
       printf("   Connection timeout: %i\n", ws->settings.timeout);
-      printf("   Chunk size: %lu\n", ws->settings.maxdatasize);
+      printf("   Chunk size: %zu\n", ws->settings.maxdatasize);
       printf("   Callbacks set:");
       if (ws->settings.on_connect) printf(" connect");
       if (ws->settings.on_receive) printf(" receive");
       if (ws->settings.on_disconnect) printf(" disconnect");
       printf("\n");
       ll_count(iter, ws->conns, conns);
-      printf("   Connections: %lu\n", conns);
+      printf("   Connections: %zu\n", conns);
       printf("   Socket descriptor: %i\n", ws->sockfd);
    }
 }
@@ -106,7 +106,7 @@ void ws_conn_print(struct ws_conn *conn)
       printf("   Timeout: ");
       if (conn->timeout) printf("Active\n");
       else printf("Deactive\n");
-      printf("   Data waiting (%lu chars): '", conn->send_len);
+      printf("   Data waiting (%zu chars): '", conn->send_len);
       for (i = 0; i < conn->send_len; i++) {
          if (conn->send_msg[i] == '\\') printf("\\\\");
          else if (isprint(conn->send_msg[i])) printf("%c", conn->send_msg[i]);
